fix : small changes

+++ src/main/java/mate/academy/bookstore/serivce/impl/BookServiceImpl.java
@@ -50,6 +50,9 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public void deleteById(Long id) {
+        if (!bookRepository.existsById(id)) {
+            throw new EntityNotFoundException("Can't delete book by id: " + id);
+        }
         bookRepository.deleteById(id);
     }
 }
